added app config
