+++ songles/App.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<configuration>
+  <appSettings>
+    <add key="Debug" value="True"/>
+    <add key="SonglesDb" value="songles.db" />
+    <add key="SonglesCSVPath" value="Data/songles.csv" />
+  </appSettings>
+</configuration>
